# Open-API/Dashboard-OpenAPI.yml
openapi: 3.0.3
info:
  title: Immersive-Dashboard
  version: 1.0.0
  description: Dashboard for ALTERRA Academy - Immersive Program
  contact:
    name: Official
    url: https://academy.alterra.id
    email: Academy@alterra.id
...
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/FEBRYANZAINAL/Immersive-Dashboard-OpenAPI/1.0.0
  - description: Real Immersive-Dashboard servers 
    url: http:// 34.87.101.252:80
    variables:
      environment:
        default: dev
        enum:
          - "dev"
          - "prod"
...
paths:
...
  /mentees:
    post:
      tags:
        - Mentees
      summary: add mentee
      description: Menambah data Mentee (register)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                mentee_name: "Mahmuda Karima"
                class_id: 1
                email: "daka@gmail.com"
                phone: "081234567"
                current_address: "jl rumah saya palembang"
                home_address: "jl rumah saya palembang"
                telegram: "@daka"
                gender: "L"
                status: "interview"
                education_type: "non-informatics"
                education_major: "akuntansi"
                education_graduate: "2022"
                education_institution: "SMA 1 Palembang"
                emergency_name: "bapak ku"
                emergency_phone: "08123457"
                emergency_status: "orang tua"
      responses:
        '201':
          description: Register berhasil.
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "register berhasil"
        '400':
          description: Register gagal.
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "failed"
                  message: "register gagal. semua field wajib diisi"
        '500':
          $ref: '#/components/responses/500'
          #description: Register gagal.
          #content:
            #application/json:
              #schema:
                #type: object
                #example:
                  #status: "failed"
                  #message: "register gagal"
    get:
      tags:
        - Mentees
      summary: Get All Mentee
      description: Menampilkan semua data Mentee 
      security:
         - JWTAuth: []
      parameters:
        - name: limit
          description: "limit"
          required: false
          in: query
          schema:
            type: integer
            example: 10
        - name: offset
          description: "offset"
          required: false
          in: query
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: A list of all mentees
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success get all mentees"
                  data:
                    - id: 1
                      mentee_name: "M.Yusuf"
                      class:
                        {
                          id: 1,
                          class_name : "BE 11"
                        }
                      status: "On Class"
                      education_type: "non-informatics"
                      gender : "L"

                    - id: 2
                      mentee_name: "Wendy"
                      class:
                        {
                          id: 2,
                          class_name : "FE_8"
                        }
                      status: "On Class"
                      education_type: "non-informatics"
                      gender : "L"
                  
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/schemas/ResponseWithoutData'
  /mentees/{id_mentee}:
    get:
      tags:
        - Mentees
      summary: get mentees by id
      description: melihat data mentees by id
      security:
        - JWTAuth: []
      parameters:
        - name: id_mentee
          description: "id mentee"
          required: true
          in: path
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: list of mentees by id
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success get mentees by id"
                  data:
                    - id: 1
                      mentee_name: "M. Yusup"
                      class:
                        {
                          id: 1,
                          class_name : "BE_11"
                        }
                      email: "myusup@gmail.com"
                      phone: "081234567"
                      current_address: "jl rumah saya palembang"
                      home_address: "jl rumah saya palembang"
                      telegram: "@daka"
                      gender: "L"
                      education_type: "non-informatics"
                      education_major: "akuntansi"
                      education_graduate: "2022"
                      education_institution: "SMA 1 Palembang"
                      emergency_name: "bapak ku"
                      emergency_phone: "08123457"
                      emergency_status: "orang tua"                 
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Mentees
      summary: delete mentees by id
      description: delete data mentees by id
      security:
        - JWTAuth: []
      parameters:
        - name: id_mentee
          description: "id mentees"
          required: true
          in: path
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: delete mentees status
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success delete mentees"
        '404':
          $ref: '#/components/responses/404'
    put:
      tags:
        - Mentees
      summary: update specific mentee by id
      description: update specific data mentees by id
      security:
        - JWTAuth: []
      parameters:
        - name: id_mentee
          description: "id mentees"
          required: true
          in: path
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                - id: 1
                  mentee_name: "M. Yusup"
                  class_id: 1
                  email: "myusup@gmail.com"
                  phone: "081234567"
                  current_address: "jl rumah saya palembang"
                  home_address: "jl rumah saya palembang"
                  telegram: "@daka"
                  gender: "L"
                  education_type: "non-informatics"
                  education_major: "akuntansi"
                  education_graduate: "2022"
                  education_institution: "SMA 1 Palembang"
                  emergency_name: "bapak ku"
                  emergency_phone: "08123457"
                  emergency_status: "orang tua"
      responses:
        '200':
          description: A list of Mentees
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success update mentees"
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"
  schemas:
...
    ResponseWithoutData:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
  responses:
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
          example:
            status: "failed"
            message: "data not found"
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
          example:
            status: "failed"
            message: "internal server error"